backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
argcomplete==2.0.0
async-timeout==4.0.2
attrs==22.1.0
autoflake==1.7.7
bidict==0.22.0
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
flake8==5.0.4
frozenlist==1.3.1
idna==3.4
mccabe==0.7.0
multidict==6.0.2
mypy==0.982
mypy-extensions==0.4.3
packaging==21.3
pipx==1.1.0
pycodestyle==2.9.1
pyflakes==2.5.0
pymongo==4.3.2
pyparsing==3.0.9
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
redis==4.3.4
tomli==2.0.1
typing_extensions==4.4.0
userpath==1.8.0
wrapt==1.14.1
yarl==1.8.1
